Fix index out of bounds error
The new OptionsGrid() function added to Rook.cs results in an index out of bounds error.

Make sure both rank and file are set correctly prior to entering a while loop.
This includes restoring file/rank to currFile/currRank if the value of file/rank has previously been changed and is not altered within the upcoming while loop.