Fix OptionsGrid() call for queens in ValidDestinations()
Under the case for queens in Manager.cs's ValidDestinations() function, the code attempts to obtain the **King.cs** script-component from the piece.

This obviously results in a null reference error, since the piece being queried is a queen and thus does not have the King.cs script as a component.

Change the function call so that it instead requests the **Queen.cs** script-component.